Extracts ABS series metadata directly from Excel spreadsheets and converts to long-form.

Extracts ABS series metadata directly from Excel spreadsheets and converts to long-form.

read_abs_metadata(path, sheet)

Arguments

path
Filepath to Excel spreadsheet.

sheet
Sheet name or number.

Value

Long-form dataframe

Examples

data <- system.file("extdata","5206002_expenditure_volume_measures.xls", package = "readabs")
read_abs_metadata(path= data, sheet=2)

#>                                                                                                                                           NA
#> 1                                                                 General government - National ;  Final consumption expenditure - Defence ;
#> 2                                                             General government - National ;  Final consumption expenditure - Non-defence ;
#> 3                                                                           General government - National ;  Final consumption expenditure ;
#> 4                                                                    General government - State and local ;  Final consumption expenditure ;
#> 5                                                                                      General government ;  Final consumption expenditure ;
#> 6                                                                                              Households ;  Final consumption expenditure ;
#> 7                                                                                             All sectors ;  Final consumption expenditure ;
#> 8                                                                      Private ;  Gross fixed capital formation - Dwellings - New and Used ;
#> 9                                                         Private ;  Gross fixed capital formation - Dwellings - Alterations and additions ;
#> 10                                                                            Private ;  Gross fixed capital formation - Dwellings - Total ;
#> 11                                                                     Private ;  Gross fixed capital formation - Ownership transfer costs ;
#> 12                                                     Private ;  Gross fixed capital formation - Non-dwelling construction - New building ;
#> 13                                     Private ;  Gross fixed capital formation - Non-dwelling construction - New engineering construction ;
#> 14                                                            Private ;  Gross fixed capital formation - Non-dwelling construction - Total ;
#> 15                                                                Private ;  Gross fixed capital formation - Machinery and equipment - New ;
#> 16                                                              Private ;  Gross fixed capital formation - Machinery and equipment - Total ;
#> 17                                                              Private ;  Gross fixed capital formation - Cultivated biological resources ;
#> 18                                    Private ;  Gross fixed capital formation - Intellectual property products - Research and development ;
#> 19                           Private ;  Gross fixed capital formation - Intellectual property products - Mineral and petroleum exploration ;
#> 20                                           Private ;  Gross fixed capital formation - Intellectual property products - Computer software ;
#> 21                                          Private ;  Gross fixed capital formation - Intellectual property products - Artistic originals ;
#> 22                                                               Private ;  Gross fixed capital formation - Intellectual property products ;
#> 23                                                            Private ;  Gross fixed capital formation - Total private business investment ;
#> 24                                                                                                Private ;  Gross fixed capital formation ;
#> 25                                                                     Public corporations - Commonwealth ;  Gross fixed capital formation ;
#> 26                                                                  Public corporations - State and local ;  Gross fixed capital formation ;
#> 27                                                                                    Public corporations ;  Gross fixed capital formation ;
#> 28                                                                General government - National ;  Gross fixed capital formation - Defence ;
#> 29                                                            General government - National ;  Gross fixed capital formation - Non-defence ;
#> 30                                                                          General government - National ;  Gross fixed capital formation ;
#> 31                                                                   General government - State and local ;  Gross fixed capital formation ;
#> 32                                                                                     General government ;  Gross fixed capital formation ;
#> 33                                                                                                 Public ;  Gross fixed capital formation ;
#> 34                                                                                            All sectors ;  Gross fixed capital formation ;
#> 35                                                                                                                   Domestic final demand ;
#> 36                                                                                                                  Changes in inventories ;
#> 37                                                                                                              Gross national expenditure ;
#> 38                                                                                                           Exports of goods and services ;
#> 39                                                                                                           Imports of goods and services ;
#> 40                                                                                                             Statistical discrepancy (E) ;
#> 41                                                                                                                  GROSS DOMESTIC PRODUCT ;
#> 42                                            General government - National ;  Final consumption expenditure - Defence: Percentage changes ;
#> 43                                        General government - National ;  Final consumption expenditure - Non-defence: Percentage changes ;
#> 44                                                      General government - National ;  Final consumption expenditure: Percentage changes ;
#> 45                                               General government - State and local ;  Final consumption expenditure: Percentage changes ;
#> 46                                                                 General government ;  Final consumption expenditure: Percentage changes ;
#> 47                                                                         Households ;  Final consumption expenditure: Percentage changes ;
#> 48                                                                        All sectors ;  Final consumption expenditure: Percentage changes ;
#> 49                                                 Private ;  Gross fixed capital formation - Dwellings - New and Used: Percentage changes ;
#> 50                                    Private ;  Gross fixed capital formation - Dwellings - Alterations and additions: Percentage changes ;
#> 51                                                        Private ;  Gross fixed capital formation - Dwellings - Total: Percentage changes ;
#> 52                                                 Private ;  Gross fixed capital formation - Ownership transfer costs: Percentage changes ;
#> 53                                 Private ;  Gross fixed capital formation - Non-dwelling construction - New building: Percentage changes ;
#> 54                 Private ;  Gross fixed capital formation - Non-dwelling construction - New engineering construction: Percentage changes ;
#> 55                                        Private ;  Gross fixed capital formation - Non-dwelling construction - Total: Percentage changes ;
#> 56                                            Private ;  Gross fixed capital formation - Machinery and equipment - New: Percentage changes ;
#> 57                                          Private ;  Gross fixed capital formation - Machinery and equipment - Total: Percentage changes ;
#> 58                                          Private ;  Gross fixed capital formation - Cultivated biological resources: Percentage changes ;
#> 59                Private ;  Gross fixed capital formation - Intellectual property products - Research and development: Percentage changes ;
#> 60       Private ;  Gross fixed capital formation - Intellectual property products - Mineral and petroleum exploration: Percentage changes ;
#> 61                       Private ;  Gross fixed capital formation - Intellectual property products - Computer software: Percentage changes ;
#> 62                      Private ;  Gross fixed capital formation - Intellectual property products - Artistic originals: Percentage changes ;
#> 63                                           Private ;  Gross fixed capital formation - Intellectual property products: Percentage changes ;
#> 64                                        Private ;  Gross fixed capital formation - Total private business investment: Percentage changes ;
#> 65                                                                            Private ;  Gross fixed capital formation: Percentage changes ;
#> 66                                                 Public corporations - Commonwealth ;  Gross fixed capital formation: Percentage changes ;
#> 67                                              Public corporations - State and local ;  Gross fixed capital formation: Percentage changes ;
#> 68                                                                Public corporations ;  Gross fixed capital formation: Percentage changes ;
#> 69                                            General government - National ;  Gross fixed capital formation - Defence: Percentage changes ;
#> 70                                        General government - National ;  Gross fixed capital formation - Non-defence: Percentage changes ;
#> 71                                                      General government - National ;  Gross fixed capital formation: Percentage changes ;
#> 72                                               General government - State and local ;  Gross fixed capital formation: Percentage changes ;
#> 73                                                                 General government ;  Gross fixed capital formation: Percentage changes ;
#> 74                                                                             Public ;  Gross fixed capital formation: Percentage changes ;
#> 75                                                                        All sectors ;  Gross fixed capital formation: Percentage changes ;
#> 76                                                                                               Domestic final demand: Percentage changes ;
#> 77                                                                                          Gross national expenditure: Percentage changes ;
#> 78                                                                                       Exports of goods and services: Percentage changes ;
#> 79                                                                                       Imports of goods and services: Percentage changes ;
#> 80                                                                                              GROSS DOMESTIC PRODUCT: Percentage changes ;
#> 81                                                                General government - National ;  Final consumption expenditure - Defence ;
#> 82                                                            General government - National ;  Final consumption expenditure - Non-defence ;
#> 83                                                                          General government - National ;  Final consumption expenditure ;
#> 84                                                                   General government - State and local ;  Final consumption expenditure ;
#> 85                                                                                     General government ;  Final consumption expenditure ;
#> 86                                                                                             Households ;  Final consumption expenditure ;
#> 87                                                                                            All sectors ;  Final consumption expenditure ;
#> 88                                                                     Private ;  Gross fixed capital formation - Dwellings - New and Used ;
#> 89                                                        Private ;  Gross fixed capital formation - Dwellings - Alterations and additions ;
#> 90                                                                            Private ;  Gross fixed capital formation - Dwellings - Total ;
#> 91                                                                     Private ;  Gross fixed capital formation - Ownership transfer costs ;
#> 92                                                     Private ;  Gross fixed capital formation - Non-dwelling construction - New building ;
#> 93                                     Private ;  Gross fixed capital formation - Non-dwelling construction - New engineering construction ;
#> 94                               Private ;  Gross fixed capital formation - Non-dwelling construction - Net purchase of second hand assets ;
#> 95                                                            Private ;  Gross fixed capital formation - Non-dwelling construction - Total ;
#> 96                                                                Private ;  Gross fixed capital formation - Machinery and equipment - New ;
#> 97                                 Private ;  Gross fixed capital formation - Machinery and equipment - Net purchase of second hand assets ;
#> 98                                                              Private ;  Gross fixed capital formation - Machinery and equipment - Total ;
#> 99                                                              Private ;  Gross fixed capital formation - Cultivated biological resources ;
#> 100                                   Private ;  Gross fixed capital formation - Intellectual property products - Research and development ;
#> 101                          Private ;  Gross fixed capital formation - Intellectual property products - Mineral and petroleum exploration ;
#> 102                                          Private ;  Gross fixed capital formation - Intellectual property products - Computer software ;
#> 103                                         Private ;  Gross fixed capital formation - Intellectual property products - Artistic originals ;
#> 104                                                              Private ;  Gross fixed capital formation - Intellectual property products ;
#> 105                                                           Private ;  Gross fixed capital formation - Total private business investment ;
#> 106                                                                                               Private ;  Gross fixed capital formation ;
#> 107                                                                    Public corporations - Commonwealth ;  Gross fixed capital formation ;
#> 108                                                                 Public corporations - State and local ;  Gross fixed capital formation ;
#> 109                                                                                   Public corporations ;  Gross fixed capital formation ;
#> 110                                                               General government - National ;  Gross fixed capital formation - Defence ;
#> 111                                                           General government - National ;  Gross fixed capital formation - Non-defence ;
#>             Unit         Series Type Data Type Frequency Collection Month
#> 1     $ Millions               Trend   DERIVED   Quarter                3
#> 2     $ Millions               Trend   DERIVED   Quarter                3
#> 3     $ Millions               Trend   DERIVED   Quarter                3
#> 4     $ Millions               Trend   DERIVED   Quarter                3
#> 5     $ Millions               Trend   DERIVED   Quarter                3
#> 6     $ Millions               Trend   DERIVED   Quarter                3
#> 7     $ Millions               Trend   DERIVED   Quarter                3
#> 8     $ Millions               Trend   DERIVED   Quarter                3
#> 9     $ Millions               Trend   DERIVED   Quarter                3
#> 10    $ Millions               Trend   DERIVED   Quarter                3
#> 11    $ Millions               Trend   DERIVED   Quarter                3
#> 12    $ Millions               Trend   DERIVED   Quarter                3
#> 13    $ Millions               Trend   DERIVED   Quarter                3
#> 14    $ Millions               Trend   DERIVED   Quarter                3
#> 15    $ Millions               Trend   DERIVED   Quarter                3
#> 16    $ Millions               Trend   DERIVED   Quarter                3
#> 17    $ Millions               Trend   DERIVED   Quarter                3
#> 18    $ Millions               Trend   DERIVED   Quarter                3
#> 19    $ Millions               Trend   DERIVED   Quarter                3
#> 20    $ Millions               Trend   DERIVED   Quarter                3
#> 21    $ Millions               Trend   DERIVED   Quarter                3
#> 22    $ Millions               Trend   DERIVED   Quarter                3
#> 23    $ Millions               Trend   DERIVED   Quarter                3
#> 24    $ Millions               Trend   DERIVED   Quarter                3
#> 25    $ Millions               Trend   DERIVED   Quarter                3
#> 26    $ Millions               Trend   DERIVED   Quarter                3
#> 27    $ Millions               Trend   DERIVED   Quarter                3
#> 28    $ Millions               Trend   DERIVED   Quarter                3
#> 29    $ Millions               Trend   DERIVED   Quarter                3
#> 30    $ Millions               Trend   DERIVED   Quarter                3
#> 31    $ Millions               Trend   DERIVED   Quarter                3
#> 32    $ Millions               Trend   DERIVED   Quarter                3
#> 33    $ Millions               Trend   DERIVED   Quarter                3
#> 34    $ Millions               Trend   DERIVED   Quarter                3
#> 35    $ Millions               Trend   DERIVED   Quarter                3
#> 36    $ Millions               Trend   DERIVED   Quarter                3
#> 37    $ Millions               Trend   DERIVED   Quarter                3
#> 38    $ Millions               Trend   DERIVED   Quarter                3
#> 39    $ Millions               Trend   DERIVED   Quarter                3
#> 40    $ Millions               Trend   DERIVED   Quarter                3
#> 41    $ Millions               Trend   DERIVED   Quarter                3
#> 42       Percent               Trend   DERIVED   Quarter                3
#> 43       Percent               Trend   DERIVED   Quarter                3
#> 44       Percent               Trend   DERIVED   Quarter                3
#> 45       Percent               Trend   DERIVED   Quarter                3
#> 46       Percent               Trend   DERIVED   Quarter                3
#> 47       Percent               Trend   DERIVED   Quarter                3
#> 48       Percent               Trend   DERIVED   Quarter                3
#> 49       Percent               Trend   DERIVED   Quarter                3
#> 50       Percent               Trend   DERIVED   Quarter                3
#> 51       Percent               Trend   DERIVED   Quarter                3
#> 52       Percent               Trend   DERIVED   Quarter                3
#> 53       Percent               Trend   DERIVED   Quarter                3
#> 54       Percent               Trend   DERIVED   Quarter                3
#> 55       Percent               Trend   DERIVED   Quarter                3
#> 56       Percent               Trend   DERIVED   Quarter                3
#> 57       Percent               Trend   DERIVED   Quarter                3
#> 58       Percent               Trend   DERIVED   Quarter                3
#> 59       Percent               Trend   DERIVED   Quarter                3
#> 60       Percent               Trend   DERIVED   Quarter                3
#> 61       Percent               Trend   DERIVED   Quarter                3
#> 62       Percent               Trend   DERIVED   Quarter                3
#> 63       Percent               Trend   DERIVED   Quarter                3
#> 64       Percent               Trend   DERIVED   Quarter                3
#> 65       Percent               Trend   DERIVED   Quarter                3
#> 66       Percent               Trend   DERIVED   Quarter                3
#> 67       Percent               Trend   DERIVED   Quarter                3
#> 68       Percent               Trend   DERIVED   Quarter                3
#> 69       Percent               Trend   DERIVED   Quarter                3
#> 70       Percent               Trend   DERIVED   Quarter                3
#> 71       Percent               Trend   DERIVED   Quarter                3
#> 72       Percent               Trend   DERIVED   Quarter                3
#> 73       Percent               Trend   DERIVED   Quarter                3
#> 74       Percent               Trend   DERIVED   Quarter                3
#> 75       Percent               Trend   DERIVED   Quarter                3
#> 76       Percent               Trend   DERIVED   Quarter                3
#> 77       Percent               Trend   DERIVED   Quarter                3
#> 78       Percent               Trend   DERIVED   Quarter                3
#> 79       Percent               Trend   DERIVED   Quarter                3
#> 80       Percent               Trend   DERIVED   Quarter                3
#> 81    $ Millions Seasonally Adjusted   DERIVED   Quarter                3
#> 82    $ Millions Seasonally Adjusted   DERIVED   Quarter                3
#> 83    $ Millions Seasonally Adjusted   DERIVED   Quarter                3
#> 84    $ Millions Seasonally Adjusted   DERIVED   Quarter                3
#> 85    $ Millions Seasonally Adjusted   DERIVED   Quarter                3
#> 86    $ Millions Seasonally Adjusted   DERIVED   Quarter                3
#> 87    $ Millions Seasonally Adjusted   DERIVED   Quarter                3
#> 88    $ Millions Seasonally Adjusted   DERIVED   Quarter                3
#> 89    $ Millions Seasonally Adjusted   DERIVED   Quarter                3
#> 90    $ Millions Seasonally Adjusted   DERIVED   Quarter                3
#> 91    $ Millions Seasonally Adjusted   DERIVED   Quarter                3
#> 92    $ Millions Seasonally Adjusted   DERIVED   Quarter                3
#> 93    $ Millions Seasonally Adjusted   DERIVED   Quarter                3
#> 94    $ Millions Seasonally Adjusted   DERIVED   Quarter                3
#> 95    $ Millions Seasonally Adjusted   DERIVED   Quarter                3
#> 96    $ Millions Seasonally Adjusted   DERIVED   Quarter                3
#> 97    $ Millions Seasonally Adjusted   DERIVED   Quarter                3
#> 98    $ Millions Seasonally Adjusted   DERIVED   Quarter                3
#> 99    $ Millions Seasonally Adjusted   DERIVED   Quarter                3
#> 100   $ Millions Seasonally Adjusted   DERIVED   Quarter                3
#> 101   $ Millions Seasonally Adjusted   DERIVED   Quarter                3
#> 102   $ Millions Seasonally Adjusted   DERIVED   Quarter                3
#> 103   $ Millions Seasonally Adjusted   DERIVED   Quarter                3
#> 104   $ Millions Seasonally Adjusted   DERIVED   Quarter                3
#> 105   $ Millions Seasonally Adjusted   DERIVED   Quarter                3
#> 106   $ Millions Seasonally Adjusted   DERIVED   Quarter                3
#> 107   $ Millions Seasonally Adjusted   DERIVED   Quarter                3
#> 108   $ Millions Seasonally Adjusted   DERIVED   Quarter                3
#> 109   $ Millions Seasonally Adjusted   DERIVED   Quarter                3
#> 110   $ Millions Seasonally Adjusted   DERIVED   Quarter                3
#> 111   $ Millions Seasonally Adjusted   DERIVED   Quarter                3
#>     Series Start Series End No. Obs
#> 1     1959-09-01 2017-03-01     231
#> 2     1959-09-01 2017-03-01     231
#> 3     1959-09-01 2017-03-01     231
#> 4     1959-09-01 2017-03-01     231
#> 5     1959-09-01 2017-03-01     231
#> 6     1959-09-01 2017-03-01     231
#> 7     1959-09-01 2017-03-01     231
#> 8     1959-09-01 2017-03-01     231
#> 9     1959-09-01 2017-03-01     231
#> 10    1959-09-01 2017-03-01     231
#> 11    1959-09-01 2017-03-01     231
#> 12    1976-09-01 2017-03-01     163
#> 13    1976-09-01 2017-03-01     163
#> 14    1959-09-01 2017-03-01     231
#> 15    1959-09-01 2017-03-01     231
#> 16    1959-09-01 2017-03-01     231
#> 17    1985-09-01 2017-03-01     127
#> 18    1985-09-01 2017-03-01     127
#> 19    1985-09-01 2017-03-01     127
#> 20    1985-09-01 2017-03-01     127
#> 21    1985-09-01 2017-03-01     127
#> 22    1959-09-01 2017-03-01     231
#> 23    1985-09-01 2017-03-01     127
#> 24    1959-09-01 2017-03-01     231
#> 25    1959-09-01 2017-03-01     231
#> 26    1959-09-01 2017-03-01     231
#> 27    1959-09-01 2017-03-01     231
#> 28    1959-09-01 2017-03-01     231
#> 29    1959-09-01 2017-03-01     231
#> 30    1959-09-01 2017-03-01     231
#> 31    1959-09-01 2017-03-01     231
#> 32    1959-09-01 2017-03-01     231
#> 33    1959-09-01 2017-03-01     231
#> 34    1959-09-01 2017-03-01     231
#> 35    1959-09-01 2017-03-01     231
#> 36    1974-09-01 2017-03-01     171
#> 37    1959-09-01 2017-03-01     231
#> 38    1959-09-01 2017-03-01     231
#> 39    1959-09-01 2017-03-01     231
#> 40    1959-09-01 2017-03-01     231
#> 41    1959-09-01 2017-03-01     231
#> 42    1959-12-01 2017-03-01     230
#> 43    1959-12-01 2017-03-01     230
#> 44    1959-12-01 2017-03-01     230
#> 45    1959-12-01 2017-03-01     230
#> 46    1959-12-01 2017-03-01     230
#> 47    1959-12-01 2017-03-01     230
#> 48    1959-12-01 2017-03-01     230
#> 49    1959-12-01 2017-03-01     230
#> 50    1959-12-01 2017-03-01     230
#> 51    1959-12-01 2017-03-01     230
#> 52    1959-12-01 2017-03-01     230
#> 53    1976-12-01 2017-03-01     162
#> 54    1976-12-01 2017-03-01     162
#> 55    1959-12-01 2017-03-01     230
#> 56    1959-12-01 2017-03-01     230
#> 57    1959-12-01 2017-03-01     230
#> 58    1985-12-01 2017-03-01     126
#> 59    1985-12-01 2017-03-01     126
#> 60    1985-12-01 2017-03-01     126
#> 61    1985-12-01 2017-03-01     126
#> 62    1985-12-01 2017-03-01     126
#> 63    1959-12-01 2017-03-01     230
#> 64    1985-12-01 2017-03-01     126
#> 65    1959-12-01 2017-03-01     230
#> 66    1959-12-01 2017-03-01     230
#> 67    1959-12-01 2017-03-01     230
#> 68    1959-12-01 2017-03-01     230
#> 69    1959-12-01 2017-03-01     230
#> 70    1959-12-01 2017-03-01     230
#> 71    1959-12-01 2017-03-01     230
#> 72    1959-12-01 2017-03-01     230
#> 73    1959-12-01 2017-03-01     230
#> 74    1959-12-01 2017-03-01     230
#> 75    1959-12-01 2017-03-01     230
#> 76    1959-12-01 2017-03-01     230
#> 77    1959-12-01 2017-03-01     230
#> 78    1959-12-01 2017-03-01     230
#> 79    1959-12-01 2017-03-01     230
#> 80    1959-12-01 2017-03-01     230
#> 81    1959-09-01 2017-03-01     231
#> 82    1959-09-01 2017-03-01     231
#> 83    1959-09-01 2017-03-01     231
#> 84    1959-09-01 2017-03-01     231
#> 85    1959-09-01 2017-03-01     231
#> 86    1959-09-01 2017-03-01     231
#> 87    1959-09-01 2017-03-01     231
#> 88    1959-09-01 2017-03-01     231
#> 89    1959-09-01 2017-03-01     231
#> 90    1959-09-01 2017-03-01     231
#> 91    1959-09-01 2017-03-01     231
#> 92    1976-09-01 2017-03-01     163
#> 93    1976-09-01 2017-03-01     163
#> 94    1959-09-01 2017-03-01     231
#> 95    1959-09-01 2017-03-01     231
#> 96    1959-09-01 2017-03-01     231
#> 97    1970-09-01 2017-03-01     187
#> 98    1959-09-01 2017-03-01     231
#> 99    1985-09-01 2017-03-01     127
#> 100   1985-09-01 2017-03-01     127
#> 101   1985-09-01 2017-03-01     127
#> 102   1985-09-01 2017-03-01     127
#> 103   1985-09-01 2017-03-01     127
#> 104   1959-09-01 2017-03-01     231
#> 105   1985-09-01 2017-03-01     127
#> 106   1959-09-01 2017-03-01     231
#> 107   1959-09-01 2017-03-01     231
#> 108   1959-09-01 2017-03-01     231
#> 109   1959-09-01 2017-03-01     231
#> 110   1959-09-01 2017-03-01     231
#> 111   1959-09-01 2017-03-01     231
#>  [ reached getOption("max.print") -- omitted 139 rows ]