read_awe

Convenience function to obtain wage levels from ABS 6302.0, Average Weekly Earnings, Australia.

read_awe(
  wage_measure = c("awote", "ftawe", "awe"),
  sex = c("persons", "males", "females"),
  sector = c("total", "private", "public"),
  state = c("all", "nsw", "vic", "qld", "sa", "wa", "tas", "nt", "act"),
  na.rm = FALSE,
  path = Sys.getenv("R_READABS_PATH", unset = tempdir()),
  show_progress_bars = FALSE,
  check_local = FALSE
)

Arguments

wage_measure
Character of length 1. Must be one of:

`awote` Average weekly ordinary time earnings; also known as Full-time adult ordinary time earnings

`ftawe` Full-time adult total earnings

`awe` Average weekly total earnings of all employees

sex
Character of length 1. Must be one of: `persons`, `males`, or `females`.

sector
Character of length 1. Must be one of: `total`, `private`, or `public`. Note that you cannot get sector-by-state data; if `state` is not `all` then `sector` must be `total`.

state
Character of length 1. Must be one of: `all`, `nsw`, `vic`, `qld`, `sa`, `wa`, `nt`, or `act`. Note that you cannot get sector-by-state data; if `sector` is not `total` then `state` must be `all`.

na.rm
Logical. `FALSE` by default. If `FALSE`, a consistent quarterly series is returned, with `NA` values for quarters in which there is no data. If `TRUE`, only dates with data are included in the returned data frame.

path
See `?read_abs`

show_progress_bars
See `?read_abs`

check_local
See `?read_abs`

Value

A `tbl_df` with four columns: `date`, `sex`, `wage_measure` and `value`. The data is nominal and seasonally adjusted.

Details

The latest AWE data is available using `read_abs(cat_no = "6302.0", tables = 2)`. However, this time series only goes back to 2012, when the ABS switched from quarterly to biannual collection and release of the AWE data. The `read_awe()` function assembles on time series back to November 1983 quarter; it is quarterly to 2012 and biannual from then. Note that the data returned with this function is consistently quarterly; any quarters for which there are no observations are recorded as `NA` unless `na.rm` = `TRUE`.

Examples

if (FALSE) {
read_awe("awote", "persons")
}