Download a tidy tibble containing the Consumer Price Index from the ABS

read_cpi() uses the read_abs() function to download, import, and tidy the quarterly Consumer Price Index from the ABS. It returns a tibble containing two columns: the date and the CPI index value that corresponds to that date. This makes joining the CPI to another dataframe easy. read_cpi() returns the original (ie. not seasonally adjusted) all groups CPI for Australia. If you want the analytical series (eg. seasonally adjusted CPI, or trimmed mean CPI) or monthly series, you can use read_abs().

read_cpi(
  frequency = c("quarterly", "monthly"),
  path = Sys.getenv("R_READABS_PATH", unset = tempdir()),
  ...
)

Arguments

frequency

Either "quarterly" (the default) or "monthly".

path

character; default is "data/ABS". Only used if retain_files is set to TRUE. Local directory in which to save downloaded ABS time series spreadsheets.

...

Arguments passed to read_abs(). See ?read_abs for details.

Details

By default, read_abs() retrieves the quarterly CPI. It can use the frequency argument to change to monthly.

Examples

# Create a tibble called 'cpi' that contains the CPI index
# numbers for each quarter
# \donttest{
cpi <- read_cpi()
#> Finding URLs for tables corresponding to ABS series IDA2325846C
#> Attempting to download files from series IDA2325846C, Consumer Price Index, Australia
#> Extracting data from downloaded spreadsheets
#> Tidying data from imported ABS spreadsheets
# }

# This tibble can now be joined to another to help streamline the process of
# deflating nominal values.