Convenience function to download and tidy data cubes from ABS Labour Force, Australia, Detailed.

Convenience function to download and tidy data cubes from ABS Labour Force, Australia, Detailed.

read_lfs_datacube(cube, path = Sys.getenv("R_READABS_PATH", unset = tempdir()))

Arguments

cube

character. A character string that is either the complete filename or (uniquely) in the filename of the data cube you want to download. Use get_available_lfs_cubes() to see a dataframe of options.

path

Local directory in which downloaded files should be stored.

Value

A tibble with the data from the data cube. Columns names are tidied and dates are converted to Date class.

Examples

read_lfs_datacube("EQ02")
#> File downloaded in /var/folders/bp/kmfmhnl95kx1c8x321z7twbw0000gn/T//RtmpL42qkE/EQ02.xlsx
#> # A tibble: 9,280 × 8
#>    date       sex   number_of_months_or_years_with_curr…¹ state_and_territory_…²
#>    <date>     <chr> <chr>                                 <chr>                 
#>  1 2001-05-01 Males Fewer than 3 months                   New South Wales       
#>  2 2001-05-01 Males Fewer than 3 months                   Victoria              
#>  3 2001-05-01 Males Fewer than 3 months                   Queensland            
#>  4 2001-05-01 Males Fewer than 3 months                   South Australia       
#>  5 2001-05-01 Males Fewer than 3 months                   Western Australia     
#>  6 2001-05-01 Males Fewer than 3 months                   Tasmania              
#>  7 2001-05-01 Males Fewer than 3 months                   Northern Territory    
#>  8 2001-05-01 Males Fewer than 3 months                   Australian Capital Te…
#>  9 2001-05-01 Males 3-5 months                            New South Wales       
#> 10 2001-05-01 Males 3-5 months                            Victoria              
#> # ℹ 9,270 more rows
#> # ℹ abbreviated names:
#> #   ¹​number_of_months_or_years_with_current_employer_or_business,
#> #   ²​state_and_territory_stt__asgs_2011
#> # ℹ 4 more variables: employed_full_time_000 <dbl>,
#> #   employed_part_time_000 <dbl>,
#> #   number_of_hours_actually_worked_in_all_jobs_employed_full_time_000_hours <dbl>, …