Download and tidy ABS payroll jobs and wages data

Import a tidy tibble of ABS Weekly Payrolls data.

read_payrolls(
  series = c("industry_jobs", "subindustry_jobs", "empsize_jobs", "sex_age_jobs"),
  path = Sys.getenv("R_READABS_PATH", unset = tempdir())
)

Arguments

series

Character. Must be one of:

"industry_jobs"

Payroll jobs by industry division, state, sex, and age group (Table 4)

"subindustry_jobs"

Payroll jobs by industry sub-division and industry division (Table 6)

"empsize_jobs"

Payroll jobs by size of employer (number of employees) and state (Table 7)

"sex_age_jobs

Payroll jobs by sex and age (Table 8)

The default is "industry_jobs".

path

Local directory in which downloaded ABS time series spreadsheets should be stored. By default, path takes the value set in the environment variable "R_READABS_PATH". If this variable is not set, any files downloaded by read_abs() will be stored in a temporary directory (tempdir()).

Value

A tidy (long) tbl_df. The number of columns differs based on the series.

Details

The ABS Weekly Payroll Jobs dataset is useful to analysts of the Australian labour market. It draws upon data collected by the Australian Taxation Office as part of its Single-Touch Payroll initiative and supplements the monthly Labour Force Survey. Unfortunately, the data as published by the ABS (1) is not in a standard time series spreadsheet; and (2) is messy in various ways that make it hard to read in R. This convenience function uses download_abs_data_cube() to import the payrolls data, and then tidies it up.

Note that this ABS release used to be called Weekly Payroll Jobs and Wages Australia. The total wages series were removed from this release in mid-2023 and it was renamed to Weekly Payroll Jobs. The ability to read total wages indexes using this function was therefore also removed.

Examples

if (FALSE) { # \dontrun{
# Fetch payroll jobs by industry and state (the default, "industry_jobs")
read_payrolls()

# Payroll jobs by employer size
read_payrolls("empsize_jobs")
} # }